ParallelIteratorIterator: Allow for the orchestrating thread to steal iteration work instead of blocking.
If you have N threads and N orchestrating (calls to
ParallelIteratorIterator.next()), then no progress can be made. This
change allows those calls to .next() to, if no sub-iterator is being
processed, to independently fetch that value.

When the producers are faster, this iterator is faster. When prodcers
are slower than consumers, this all gets very complicated.