Implement a method to have holes in GeoJSON output. (#6)
* Adding isCounterClockwise() function to operate on a Feature's point list.

This requires another O(n) walk of the list of points. I may
try to bundle this logic into a geometry test.

* Adding an invterval concept.

This will eventually be used for an interval tree of some sort to index
polygons in Marching Squares. The end-goal is for fast finding of
polygons near each other so as to put holes in the polygons.

* RedBlackTree2: Implemented by copying the original RedBlack tree code and conforming it.
* RedBlackTree2{,Test}: A simple test to exercise the tree.
* RedBlackTree{,2}: Remove source and remove() method bodies.

* IntervalTree.java: Comments update.
* Adding some IntervalTree tests.
* IntervalTree.java: Fix case where we want to step left but cannot because we are at a leaf and so, wrongly, step right.

The same result is computed because stepping right only adds an extra
node to check when walking backwards through the tree. That node will be
filtered out.

* RTree.java: Hacking this up.
* RTree: Adding find and delete methods.
* RTree, RTreeTest: Testing add, find, and delete methods.

Testing for findEnclosing and findEnclosed are needed.

* RTree, RTreeTest: Testing findEnclosing and findEnclosed.
* RTree, RTreeTest: Adding deleteSubtree.
* RTree: Add forEach and abortable callbacks.
* Feature, FeatureTest: Allow checking if a point is inside a feature.
* Moving the hole collation logic into VectorTile.
* RTree: Fixing findEnclosing() where it stopped checking for matches too early.
* Wire up the collateHoles() call but it does not seem to work in some random unit tests.
* SimpleGeoJson: Fix encoding bug. VectorTileTest: Add a test.