Add Copyright notice to sdsai-common source files. Other projects will follow.
This allows for easy legal use of the files by identifying the copyright holder.
The lower-bound may be later than the file actually existed. Much of this code
base was authored near 2002 but moved and adapted into this current form. Particularly,
the $Id$ fields come from a previous version of these files stored in svn.